+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [#31](https://github.com/bobozaur/sqlx-exasol/pull/31):
   - Renamed `ExaConnection::socket_addr` to `ExaConnection::server`
   - Made `feedback_interval` match between options and attributes
+  - Removed the boxing of `EtlQuery` future
 
 ## [0.8.6] - 2025-06-03